this

+++ app/src/main/java/com/ru/htccs/converter/model/base/Converter.java
@@ -2,5 +2,5 @@ package com.ru.htccs.converter.model.base;
 
 public interface Converter {
 
-    void resultConversion(double values, double dpi);
+    void resultConversion(double values, int dpi);
 }
