+++ app/src/main/java/com/ru/htccs/converter/model/converters/ConverterDP.java
@@ -5,7 +5,7 @@ import com.ru.htccs.converter.model.base.Converter;
 public class ConverterDP implements Converter {
 
     @Override
-    public void resultConversion(double values, double dpi) {
+    public void resultConversion(double values, int dpi) {
 
     }
 }
