+++ app/src/main/java/com/ru/htccs/converter/model/converters/ConverterPT.java
@@ -5,7 +5,7 @@ import com.ru.htccs.converter.model.base.Converter;
 public class ConverterPT implements Converter {
 
     @Override
-    public void resultConversion(double values, double dpi) {
+    public void resultConversion(double values, int dpi) {
 
     }
 }
